worker/requirements.txt
flask
requests
websockets
python-socketio
websocket-client>=1.5.1